pyproject.toml
[project]
name = "demo-rest"
version = "0.1.0"
description = "Simple REST API demo with FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pyjwt>=2.9",
    "bcrypt>=4.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncmy>=0.2.9",
    "cryptography>=42.0",
    "motor>=3.4",
    "aiokafka>=0.10",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
